requirement.txt
appnope==0.1.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
backports.csv==1.0.6
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
bleach==3.0.0
certifi==2018.10.15
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.1.0
Click==7.0
decorator==4.3.0
defusedxml==0.5.0
docx==0.2.4
entrypoints==0.2.3
feedparser==5.2.1
Flask==1.0.2
future==0.17.1
idna==2.7
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
jaraco.functools==1.20
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
lxml==4.2.5
MarkupSafe==1.0
mistune==0.8.3
more-itertools==4.3.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
pandocfilters==1.4.2
parso==0.3.1
pattern3==3.0.0
pdfminer.six==20181108
pdfminer3k==1.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.8.0
ply==3.11
portend==2.3
prometheus-client==0.4.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
py==1.7.0
pycryptodome==3.7.2
Pygments==2.2.0
pytest==4.0.2
python-dateutil==2.7.3
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.20.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
sortedcontainers==2.1.0
tempora==1.14
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
zc.lockfile==1.4
